+++ src/engine/types.rs
@@ -1,7 +1,7 @@
 use std::time::Duration;
 
 /// The default engine api authentication port.
-pub const DEFAULT_AUTH_PORT: u32 = 8551;
+pub const DEFAULT_AUTH_PORT: u16 = 8551;
 
 /// The ID of the static payload
 pub const STATIC_ID: u32 = 1;
